requirements.txt
asgiref==3.3.1
cycler==0.10.0
decorator==4.4.2
Django==3.1.4
django-post-office==3.5.3
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
imageio==2.9.0
jsonfield==3.1.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
tifffile==2020.12.4
